Enable bugprone-signed-char-misuse
The check flags plain char bytes sign extending into wider integers, which
turned up four places where a pixel or byte value above 0x7F went negative.

tech/pcx_file.cc read the first pixel of a scanline as *ptr while the rest
of the loop read 0xff & *++ptr, so a run starting on a pixel >= 0x80 never
compared equal on the first iteration and the RLE run was cut short.
tech/2keyfbuf.cc built a pixel index the same way and then subscripted the
256 entry IsTranslucent table with it, reading before the array for the top
half of the palette; the sibling blit routine in the same file already used
uint8_t. Both jshell.cc files ran the small icon map lookup through (char*)
when the very next line already used (unsigned char*), and td/options.cc
read palette bytes through char* into the unsigned r/g/b it passes to
Convert_RGB_To_HSV.

Where the stored values are genuinely signed or simply wider than a byte,
retype the storage instead of casting: the shadow lookup tables in both
display.cc files carry -1/-2 sentinels, and TurnTrackType::Track,
StartTrack, FireLaunch and ProneLaunch are all small numbers living in a
char for no reason. None of these are serialized -- the track table is
static const and the launch fields sit in the static type classes -- so
widening them to int does not touch save game layout.

Obfuscate() in both init.cc files is the one place where the sign extension
was load bearing, since the resulting vector is fed to a CRC. Reading the
bytes unsigned is equivalent here: the transformation uses only +, * and ^,
whose low 8 bits depend only on the low 8 bits of their operands, and only
those low 8 bits are stored back. Verified by running both versions over
200000 random buffers with identical output, so the network key is
unchanged. The short locals in that loop become int16_t while the block is
being touched.

ra/monoc.cc cast an int down to char only to assign it straight back into
the int members X, Y and Page, so those casts just go away. ra/statbtn.cc
and td/colrlist.cc keep a narrow cast at the point of use: FontHeight is a
global char with separate definitions in sdllib and vqm32, and turning
ColorListClass::Colors into a vector of int would make DynamicVectorClass
Delete(const T&) and Delete(int index) ambiguous at the one call site that
deletes by index.

+++ src/tech/2keyfbuf.cc
@@ -69,7 +69,7 @@ static void Setup_Shape_Header(int pixel_width, int pixel_height, char* src,
     int trans_count = 0;
     int x_count = pixel_width;
     do {
-      int pixel = *src;
+      int pixel = static_cast<uint8_t>(*src);
       src = src + 1;
       if (!pixel && flags & SHAPE_TRANS) {
         line_flags = BLIT_TRANSPARENT;
